+++ sequence/__init__.py
@@ -3,6 +3,7 @@ from sequence.cli.cli import execute_command_line_interface
 from sequence.sequences.integer.explicit import *
 from sequence.sequences.integer.finite import *
 from sequence.sequences.integer.periodic import *
+from sequence.sequences.integer.periodic_generalised import *
 from sequence.sequences.integer.property_defined import *
 from sequence.sequences.integer.recursive import *